Fix Prune retention policy tests failing at certain times
In these tests the expected amount of snapshots left after pruning with a retention policy
depends on whether the most recent snapshot started yesterday or not.
(i.e If 1 hour ago is yesterday it counts as 1 day ago).
So we need to check if it started yesterday and reduce the expected amount by 1.

Also removed a duplicate retention policy test.